Changed a little bit

+++ src/com/javaguru/student_igors_gergeleziu/lesson_2/level_x/Super_task_3.java
@@ -10,8 +10,9 @@ class CorrectDataTypes {
             System.out.println("Number 1 = " + numberOne);
             System.out.println("Number 2 = " + numberTwo);
 
-            float sum = numberOne + (float)numberTwo;
-            //double sum = numberOne + numberTwo;
+            double sum = numberOne + numberTwo;
+            //float sum = numberOne + (float)numberTwo;
+
 
             System.out.println("Sum = " + sum);
         }
